Refactor testing/stub.go
The Stub type was not thread safe, yet could easily be used in a
multithreaded environment. We see this in LP 1459064.

To address this all the public fields in Stub have been made private
and either accessors which return a _copy_ of the values, ie stub.Calls,
or checkers which assert the internal contents match a supplied list, have
been added.

Additionally, unused features like Stub.DefaultError have been removed. I
could not find a usage of that inside Juju -- the value of DefaultError
was always nil.