testing: pass --nojournal to testing mongodb
We disabled preallocation of data files and reduce the size of the .ns
file (whatever that is), but on startup mongodb will still try to zero
out the journal 8kb at a time.

If you don't have /tmp mounted on tmpfs your tests will be dominated by
the time to zero this backing store.

Disable the journal for testing mongodb's