Update testing.Clock to improve robustness
Adds a WaitAdvance method which ensures that there is something in
clock.waiting before actually advancing the time. Addtionally cause
Advance to log a debug message if it Advances time when clock.waiting
is empty.

More information about this can be found on the wiki: cf.
https://github.com/juju/juju/wiki/Intermittent-failures
There was also discussion on IRC about the use of alarms; cf.
https://irclogs.ubuntu.com/2016/09/30/%23juju-dev.txt

Refs: https://bugs.launchpad.net/juju/+bug/1607044