Introduces a type which can be used to mock calls
This type is designed to easily and dynamically mock out calls to methods/functions so that tests can declare expected arguments and their return values without having to define custom functions for each test. In addition, calls and return-values are logged so that when tests fail it's easy to pinpoint why.