Fix EchoQuotedArgsUnix/AssertEchoArgs
The EchoQuotedArgsUnix script writes
a file to record the command line arguments
of the process. The script was using
the basename, which only works as
expected if the command is run in the
CWD set at the time of the AssertEchoArgs
call.

Instead of this, write to the location
of the script + ".out", and then read
from the same location in AssertEchoArgs
using os/exec.LookPath. We also introduce
ReadEchoArgs to fetch the value without
checking its contents.