Don't try to truncate the local replicaset oplog.
In Mongo 3.4 it now prevents you from doing this. It doesn't seem like
it would ever be a really good idea to do, as it would arguably break
your system. (if you really were part of a replicaset, you've now lost
your ability to track what changes have and haven't been applied.)
It was semi-ok in that we are resetting the database, but since the
replicaset would still potentially be running, we really don't want to
do this.