Allow testing of DoRequest against status code
The following changes are required so we can test against the
status code along with the error of a http request.

Changing this allows us to work with go-1.12 where the changes to
the http(s) server when calling https using http now returns a 400
status code instead of malformed garbage.

See: https://go-review.googlesource.com/c/go/+/98447/