error.Is testing checker
The following makes it easier to compare errors. Currently, if
we attempt to use satisfies checker, we'll more thank likely endup
with a deprecation warning that Is prefix predicate variants from
the errors package is being used.

The solution is to make a new checker that we can move to, which has
the semantics that we want moving forward.